+++ atlasdb-client/src/main/java/com/palantir/atlasdb/schema/SweepSchema.java
@@ -20,6 +20,7 @@ import java.io.File;
 import com.google.common.base.Supplier;
 import com.google.common.base.Suppliers;
 import com.palantir.atlasdb.keyvalue.api.Namespace;
+import com.palantir.atlasdb.table.description.OptionalType;
 import com.palantir.atlasdb.table.description.Schema;
 import com.palantir.atlasdb.table.description.TableDefinition;
 import com.palantir.atlasdb.table.description.ValueType;
@@ -39,11 +40,14 @@ public enum SweepSchema implements AtlasSchema {
     private static Schema generateSchema() {
         Schema schema = new Schema("Sweep",
                 SweepSchema.class.getPackage().getName() + ".generated",
-                NAMESPACE);
+                NAMESPACE,
+                OptionalType.JAVA8);
 
         // This table tracks progress on a sweep job of a single table.
         schema.addTableDefinition("progress", new TableDefinition() {{
             javaTableName("SweepProgress");
+            tableNameIsSafeLoggable();
+            namedComponentsSafeByDefault();
             rowName();
                 // This table has at most one row.
                 rowComponent("dummy", ValueType.VAR_LONG);
@@ -68,6 +72,8 @@ public enum SweepSchema implements AtlasSchema {
         // in determining when and in which order they should be swept.
         schema.addTableDefinition("priority", new TableDefinition() {{
             javaTableName("SweepPriority");
+            tableNameIsSafeLoggable();
+            namedComponentsSafeByDefault();
             rowName();
                 rowComponent("full_table_name", ValueType.STRING);
             columns();