+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/keyvalue/impl/TableRemappingKeyValueService.java
@@ -45,6 +45,7 @@ import com.palantir.atlasdb.keyvalue.api.RowColumnRangeIterator;
 import com.palantir.atlasdb.keyvalue.api.RowResult;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
 import com.palantir.atlasdb.keyvalue.api.Value;
+import com.palantir.atlasdb.logging.KeyValueServiceArgSupplier;
 import com.palantir.common.base.ClosableIterator;
 import com.palantir.util.paging.TokenBackedBasicResultsPage;
 
@@ -438,4 +439,13 @@ public final class TableRemappingKeyValueService extends ForwardingObject implem
         return delegate().getClusterAvailabilityStatus();
     }
 
+    @Override
+    public void rehydrateLoggingArgSupplier() {
+        delegate().rehydrateLoggingArgSupplier();
+    }
+
+    @Override
+    public KeyValueServiceArgSupplier getLoggingArgSupplier() {
+        return delegate().getLoggingArgSupplier();
+    }
 }