+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/sweep/BackgroundSweeperImpl.java
@@ -26,6 +26,7 @@ import com.google.common.base.Preconditions;
 import com.google.common.base.Supplier;
 import com.palantir.atlasdb.keyvalue.api.InsufficientConsistencyException;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
+import com.palantir.atlasdb.logging.KeyValueServiceArgSupplier;
 import com.palantir.atlasdb.sweep.priority.NextTableToSweepProvider;
 import com.palantir.atlasdb.sweep.priority.NextTableToSweepProviderImpl;
 import com.palantir.atlasdb.sweep.progress.SweepProgress;
@@ -34,7 +35,6 @@ import com.palantir.atlasdb.transaction.api.TransactionTask;
 import com.palantir.common.base.Throwables;
 import com.palantir.lock.RemoteLockService;
 import com.palantir.logsafe.SafeArg;
-import com.palantir.logsafe.UnsafeArg;
 
 public final class BackgroundSweeperImpl implements BackgroundSweeper {
     private static final Logger log = LoggerFactory.getLogger(BackgroundSweeperImpl.class);
@@ -44,6 +44,7 @@ public final class BackgroundSweeperImpl implements BackgroundSweeper {
     private final Supplier<Long> sweepPauseMillis;
     private final PersistentLockManager persistentLockManager;
     private final SpecificTableSweeper specificTableSweeper;
+    private final KeyValueServiceArgSupplier argSupplier;
 
     static volatile double batchSizeMultiplier = 1.0;
 
@@ -56,20 +57,23 @@ public final class BackgroundSweeperImpl implements BackgroundSweeper {
             Supplier<Boolean> isSweepEnabled,
             Supplier<Long> sweepPauseMillis,
             PersistentLockManager persistentLockManager,
-            SpecificTableSweeper specificTableSweeper) {
+            SpecificTableSweeper specificTableSweeper,
+            KeyValueServiceArgSupplier argSupplier) {
         this.lockService = lockService;
         this.nextTableToSweepProvider = nextTableToSweepProvider;
         this.isSweepEnabled = isSweepEnabled;
         this.sweepPauseMillis = sweepPauseMillis;
         this.persistentLockManager = persistentLockManager;
         this.specificTableSweeper = specificTableSweeper;
+        this.argSupplier = argSupplier;
     }
 
     public static BackgroundSweeperImpl create(
             Supplier<Boolean> isSweepEnabled,
             Supplier<Long> sweepPauseMillis,
             PersistentLockManager persistentLockManager,
-            SpecificTableSweeper specificTableSweeper) {
+            SpecificTableSweeper specificTableSweeper,
+            KeyValueServiceArgSupplier argSupplier) {
         NextTableToSweepProvider nextTableToSweepProvider = new NextTableToSweepProviderImpl(
                 specificTableSweeper.getKvs(), specificTableSweeper.getSweepPriorityStore());
         return new BackgroundSweeperImpl(
@@ -78,7 +82,8 @@ public final class BackgroundSweeperImpl implements BackgroundSweeper {
                 isSweepEnabled,
                 sweepPauseMillis,
                 persistentLockManager,
-                specificTableSweeper);
+                specificTableSweeper,
+                argSupplier);
     }
 
     @Override
@@ -207,7 +212,8 @@ public final class BackgroundSweeperImpl implements BackgroundSweeper {
                             Optional<TableReference> nextTable = nextTableToSweepProvider.chooseNextTableToSweep(
                                     tx, specificTableSweeper.getSweepRunner().getConservativeSweepTimestamp());
                             if (nextTable.isPresent()) {
-                                log.debug("Now starting to sweep next table.", UnsafeArg.of("table name", nextTable));
+                                log.debug("Now starting to sweep next table.",
+                                        argSupplier.getTableReferenceArg("table name", nextTable.get()));
                                 return Optional.of(new TableToSweep(nextTable.get(), null));
                             } else {
                                 return Optional.empty();