+++ docs/source/release_notes/release-notes.rst
@@ -37,6 +37,27 @@ develop
 
 .. replace this with the release date
 
+.. list-table::
+    :widths: 5 40
+    :header-rows: 1
+
+    *    - Type
+         - Change
+
+    *    - |new|
+         - AtlasDB now supports specifying the safety of table names as well as row and column component names following the `palantir/safe-logging <https://github.com/palantir/safe-logging>`__ library.
+           Please consult the documentation for :ref:`Tables and Indices <tables-and-indices>` for details on how to set this up.
+           As AtlasDB regenerates its metadata on startup, changes will take effect after restarting your AtlasDB client (in particular, you do NOT need to rerender your schemas.)
+           Previously, all table names, row component names and column names were always treated as unsafe.
+           (`Pull Request 1 <https://github.com/palantir/atlasdb/pull/VSC1>`__,
+           `Pull Request 2 <https://github.com/palantir/atlasdb/pull/VSC2>`__ and
+           `Pull Request 3 <https://github.com/palantir/atlasdb/pull/VSC3>`__)
+
+    *    - |improved|
+         - The ``ProfilingKeyValueService`` and ``SpecificTableSweeper`` now log table names as safe arguments, if and only if these have been specified as safe in one's schemas.
+           (`Pull Request <https://github.com/palantir/atlasdb/pull/VSC3>`__)
+
+
 .. <<<<------------------------------------------------------------------------------------------------------------->>>>
 
 ======